+++ .agents/skills/run-system-tests/SKILL.md
@@ -25,7 +25,8 @@ This skill is about the **test harness itself** — pytest marks, fixtures, the 
 The suite lives at `tests/` (repo root) and is fully pytest-based. Configuration is in `tests/pytest.ini` and shared infrastructure in `tests/conftest.py`.
 
 - **`tests/system/`** — Docker stack integration tests. Marks: `build_docker`, `build_packages`, `liveliness`, `sensors`, `takeoff_hover_land`, `autonomy`.
-- **`tests/robot/`** and **`tests/sim/`** — Hermetic **unit** tests (`@pytest.mark.unit`). These are **thin proxy files** that re-export tests from each ROS 2 package's own `test/` directory (co-located with the source, the ROS 2 / colcon convention). The proxy pattern keeps test source next to the code it tests while making tests discoverable by `pytest tests/`.
+- **`tests/integration/`** — Cross-component tests (`integration` mark): robot container + a host-side component, no sim/GPU.
+- **Unit tests** (`@pytest.mark.unit`) — Hermetic. Source is **co-located** with each ROS 2 package in its own `test/` dir (ROS 2 / colcon convention). `tests/colcon_unit_test_packages.yaml` lists which packages have unit tests; `conftest.py` resolves each to its `test/` dir and collects the non-linter `test_*.py` under `--import-mode=importlib`.
 
 ### Unit tests vs system tests
 
@@ -34,7 +35,7 @@ The suite lives at `tests/` (repo root) and is fully pytest-based. Configuration
 | Hardware required | None — pure Python | Docker daemon, NVIDIA GPU, sim license |
 | CI workflow | `system-tests.yml` (included in `pytest tests/`) | `system-tests.yml` (GPU OpenStack VM) |
 | Trigger | Every push + PR (automatic) | PR opened, `/pytest` comment, `workflow_dispatch` |
-| Source location | `<pkg>/test/test_*.py` (proxied via `tests/robot/`) | `tests/system/` |
+| Source location | `<pkg>/test/test_*.py` (collected directly, listed in `colcon_unit_test_packages.yaml`) | `tests/system/` |
 | How to add | See `add-unit-tests` skill | See *Adding a New System Test* below |
 
 Run unit tests without any Docker stack:
@@ -45,7 +46,7 @@ airstack test -m unit -v
 pytest tests/ -m unit -v   # AIRSTACK_ROOT=$(pwd) for direct pytest
 ```
 
-For details on the proxy pattern and adding new unit tests, see the
+For details on the co-located layout and adding new unit tests, see the
 `add-unit-tests` skill.
 
 | File | Mark | What it tests | Hardware required |
@@ -80,7 +81,7 @@ rates if too many `ros2 topic hz` processes run concurrently.
 - **Robot-side (Isaac):** two passes — both stereo images, then both depths.
   **ms-airsim** keeps a single four-topic parallel batch on the robot container.
 - **Filtered LiDAR** (`PointCloud2`): uses `ros2 topic echo --once` per robot
-  (see `parallel_echo_once_robot_topics` in `conftest.py`), not `topic hz`.
+  (see `parallel_echo_once_robot_topics` in `tests/harness/sim.py`), not `topic hz`.
 - **Multi-drone Pegasus script:** pytest sets `ENABLE_LIDAR=true` in
   `conftest.py` `SIM_CONFIG["isaacsim"]["extra_env"]` so LiDAR matches the
   single-drone example (which always enables RTX LiDAR).
@@ -294,7 +295,7 @@ If your test...
 - File: `tests/system/test_<short_descriptor>.py` — matches pytest's default test discovery (`test_*.py`) under the system suite
 - Class: `Test<CamelCase>` with the mark applied at the class level: `@pytest.mark.<mark>`
 - Add a class-level `@pytest.mark.timeout(<seconds>)` — long-running sim tests need it
-- Imports: pull helpers from `conftest` directly (`from conftest import ...`); `tests/` is on `sys.path` because `testpaths = .` in pytest.ini
+- Imports: pull helpers from `conftest` directly (`from conftest import ...`); they physically live in the `tests/harness/` package but are re-exported through `conftest`, so either `from conftest import ...` or `from harness import ...` works. `tests/` is on `sys.path` because `testpaths = .` in pytest.ini
 
 ### 3. Decide if you need `airstack_env`
 
@@ -304,7 +305,7 @@ If your test...
 
 ### 4. Use the existing helpers
 
-`conftest.py` exports a deliberate API. Prefer these over rolling your own:
+The `tests/harness/` package exports a deliberate API (re-exported through `conftest`). Prefer these over rolling your own:
 
 | Helper | Purpose |
 |--------|---------|
@@ -421,7 +422,8 @@ python tests/parse_metrics.py \
 
 ### Files to know
 
-- `tests/conftest.py` — fixtures, helpers, `MetricsRecorder`, ordering hooks
+- `tests/conftest.py` — pytest hooks + the `airstack_env` / `robot_autonomy_stack` fixtures (re-exports the harness API)
+- `tests/harness/` — helpers split by concern: `session`, `discovery`, `commands`, `containers`, `metrics` (`MetricsRecorder`), `sim`, `collection` (ordering)
 - `tests/pytest.ini` — mark registration, log format
 - `tests/parse_metrics.py` — markdown reporter, regression diff
 - `tests/README.md` — user-facing docs (CLI options, output layout, CI/CD orchestrator)
