+++ tests/pytest.ini
@@ -3,12 +3,13 @@ markers =
     unit: Fast hermetic tests (no Docker stack; numpy / pure Python)
     build_docker: Docker image build tests
     build_packages: Colcon workspace build tests
+    integration: Cross-component integration tests (robot container + a host-side component; no sim/GPU)
     liveliness: Container and process health (Docker, tmux, sentinel ROS 2 nodes)
     sensors: Sim and robot sensor topic rates, LiDAR validation, sim RTF
     takeoff_hover_land: End-to-end takeoff / hover / land action tests
     autonomy: Fixed-pattern trajectory path-tracker benchmark (test_fixed_trajectory.py)
 testpaths = .
-addopts = -v --durations=0
+addopts = -v --durations=0 --import-mode=importlib
 cache_dir = /tmp/.pytest_cache
 log_cli = true
 log_cli_level = INFO
